pyproject.toml
[project]
name = "nostr-workers-relay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mecab-python3>=1.0.10",
    "requests>=2.32.4",
    "unidic-lite>=1.0.8",
    "wordcloud>=1.9.4",
]
